updated contact form and flex layout
